# backend/openAPI.yaml
openapi: 3.0.3
info:
  title: SparkLayer Coding Test
  version: 1.0.11

paths:
  /:
    post:
      summary: Add a To-Do to the list
      operationId: createToDo
      requestBody:
        description: Create a new To-Do Task
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/To-Do"
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/To-Do"
        "400":
          description: Invalid input
    get:
      summary: Gets the To-Do List
      operationId: getToDos
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/To-Do-List"
    delete:
      summary: Delete a todo item by key
      operationId: deleteTodoItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteKey'
      responses:
        '200':
          description: Updated list of todo items after deletion.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TodoItem'
        '400':
          description: Invalid key sent.
components:
  schemas:
    To-Do-List:
      additionalProperties: false
      type: array
      items:
        $ref: "#/components/schemas/To-Do"
    To-Do:
      type: object
      additionalProperties: false
      required:
        - title
        - description
      properties:
        title:
          type: string
          example: Go for a jog
        description:
          type: string
          example: run 4km around the river
    DeleteKey:
      type: object
      required:
        - key
      properties:
        key:
          type: integer
          example: 2